fix(api): Isola core Contentful no diretório de Functions
- Move handler compartilhado para app/netlify/functions para acompanhar o empacotamento

- Mantém Express local consumindo o mesmo core usado pela Function

- Adiciona teste para evitar imports fora do diretório de Functions
